requirements.txt
black
flake8
isort
loguru
mkdocs
pip
python-dotenv
tqdm
typer
pandas
numpy
seaborn
langchain
langchain-community
langchain-openai
langchain_experimental
streamlit
pygwalker
tabulate